requirements.txt
cffi==1.14.2
discord.py==1.4.1
pycparser==2.20
python-dotenv==0.14.0
PyNaCl==1.3.0
openai==0.2.6
